express-commit

+++ user_service/application/dto/Token.cs
@@ -1,7 +1,9 @@
+using user_service.services.jwt_authentification;
+
 namespace user_service.application.dto;
 
 public struct Token
 {
     public string Value { get; set; }
-    public int ExpiresSeconds { get; set; }
+    public long ExpiresSeconds { get; set; }
 }
